Upgrade get alive neighbours

+++ src/graphical_interface.rs
@@ -10,7 +10,7 @@ struct Grid {
 }
 
 /// Returns the graphical grid
-fn get_grid(squares: u32) -> Grid {
+fn get_grid(squares: usize) -> Grid {
     let game_size = screen_width().min(screen_height());
     let offset_x = (screen_width() - game_size) / 2. + 10.;
     let offset_y = (screen_height() - game_size) / 2. + 10.;
@@ -20,7 +20,7 @@ fn get_grid(squares: u32) -> Grid {
 }
 
 /// Draws the grid lines on screen
-async fn draw_grid_outline(squares: u32) {
+async fn draw_grid_outline(squares: usize) {
     let grid = get_grid(squares);
 
     for i in 0..squares+1 {
@@ -47,7 +47,7 @@ async fn draw_grid_outline(squares: u32) {
 }
 
 /// Fills a cell black given its index
-async fn fill_square(i: u32, j:u32,squares:u32 ) {
+async fn fill_square(i: u32, j:u32,squares:usize ) {
     let grid = get_grid(squares);
 
     draw_rectangle(grid.offset_x + grid.sq_size * i as f32, grid.offset_y + grid.sq_size * j as f32, grid.sq_size, grid.sq_size, BLACK);
